+++ configs/webpack.js
@@ -18,13 +18,23 @@ module.exports = function (config, mode, dirname) {
 	app.setEnactTargetsAsDefault();
 
 	const getStyleLoaders = (cssLoaderOptions = {}, preProcessor) => {
+		const mergedCssLoaderOptions = {
+			...cssLoaderOptions,
+			modules: {
+				...cssLoaderOptions.modules,
+				// Options to restore 6.x behavior:
+				// https://github.com/webpack-contrib/css-loader/blob/master/CHANGELOG.md#700-2024-04-04
+				namedExport: false,
+				exportLocalsConvention: 'as-is'
+			}
+		};
 		const loaders = [
 			process.env.INLINE_STYLES ? require.resolve('style-loader') : MiniCssExtractPlugin.loader,
 			{
 				loader: require.resolve('css-loader'),
 				options: Object.assign(
 					{importLoaders: preProcessor ? 2 : 1, sourceMap: shouldUseSourceMap},
-					cssLoaderOptions
+					mergedCssLoaderOptions
 				)
 			},
 			{
@@ -163,20 +173,6 @@ module.exports = function (config, mode, dirname) {
 			type: 'asset/resource'
 		}
 	);
-
-	// Run `source-loader` on story files to show their source code
-	// automatically in `DocsPage` or the `Source` doc block.
-	config.module.rules.push({
-		test: /\.([jt]sx?)$/,
-		loader: require.resolve('@storybook/source-loader'),
-		exclude: [/node_modules/],
-		enforce: 'pre',
-		options: {
-			injectParameters: true,
-			inspectLocalDependencies: false,
-			inspectDependencies: false
-		}
-	});
 
 	config.plugins.push(
 		new DefinePlugin({